+++ web/src/components/layout/nav/nav-logo.tsx
@@ -1,0 +1,25 @@
+import { Fragment } from 'react';
+
+import { useSession } from '../../../hooks/useSession';
+
+export const NavLogo = () => {
+  const { user } = useSession();
+  return (
+    <Fragment>
+      {user ? (
+        <div>
+          <a
+            href={`/${user.login}`}
+            className="mb-3 align-items-center mb-md-0 me-md-auto text-dark text-decoration-none"
+          >
+            <img src="/pephub_logo.svg" alt="PEPhub" height="60" />
+          </a>
+        </div>
+      ) : (
+        <a href="/" className="mb-3 align-items-center mb-md-0 me-md-auto text-dark text-decoration-none">
+          <img src="/pephub_logo.svg" alt="PEPhub" height="60" />
+        </a>
+      )}
+    </Fragment>
+  );
+};
