+++ web/src/globals.css
@@ -508,3 +508,27 @@ body {
 .translate-y-1-up {
   transform: translateY(-1rem);
 }
+
+.hidden {
+  display: none;
+}
+
+.block {
+  display: block;
+}
+
+.large-flex {
+  @media (min-width: 1024px) {
+    display: flex;
+  }
+}
+
+.large-hidden {
+  @media (min-width: 1024px) {
+    display: none;
+  }
+}
+
+.list-none {
+  list-style: none;
+}
